Request no-cache and no-store directive implementation
Section 14.9.2 - although there was implementation for no-store in a response, the request can include no-store and expect it to be obeyed as well

Section 14.9.4 - if a request includes the "no-cache" directive, then this should force an end-to-end reload. This is somewhat confusing because no-cache means almost the opposite of what it does when supplied in a response (you can cache, but must revalidate).