Fixing the intermitent problems running thin from specs.
Seems that Rspec does a lot of magic etc. and fork is "odd" within specs.
See http://everburning.com/news/forked-by-rspec/

The test server now starts up once for *every* spec, and is cleanly removed
in an at_exit hook. This has the benfit that the specs should run a bit faster
as well