Make rake spec work
simple_sinatra_server_spec and the acceptance specs cant be run in the
same go, so make the default spec runner not run them. Add a separate
rake task for those.

Also, added a rake server task that will spawn the server with some
logging, so you can watch the requests being made by the acceptance
specs. Useful for debugging.